test: app 모듈에 있던 InjectorTest를 di 모듈로 이동
- 그 과정에서 안드로이드 의존성 제거
