+++ di/build.gradle.kts
@@ -10,4 +10,13 @@ java {
 
 dependencies {
     implementation("org.jetbrains.kotlin", "kotlin-reflect", "1.8.21")
+    implementation("com.google.ar:core:1.39.0")
+    testImplementation("org.junit.jupiter", "junit-jupiter", "5.8.2")
+    testImplementation("org.assertj", "assertj-core", "3.22.0")
+}
+
+tasks {
+    test {
+        useJUnitPlatform()
+    }
 }
